backend/requirements.txt
Flask
Flask-Cors
pymongo
PyJWT
python-dotenv
